feat: add getMappedUserProfileValue utility for improved user profile attribute mapping
